Write GeneralModule, GameBoard module pattern and Player object
In the GeneralModule I've added getRandomNumber() to generate a random
number within the range of the provided number parameter.

The GameBoard module stores the currentBoardState, the markersArray
getBoardState(), which gets the current state of the board,
getMarkersArray() and  popMarker().

Finally the Player factory function will return an object with the name
and marker of the player, we will need to provide the name to the
factory function.