Start Game module pattern
In this module the logic of the tic-tac-toe game is going to be written.
As of now we've written the beginning of the game loop, that is gonna run
until the game has declared a winner.

So far we've register the names of both players, randomized which player starts
with the whoStartsGame(), captured the position in which the currentPlayer wants
to place its marker and checked with the isBoardCellFilled() if that position is
already filled.