Render marker in Board when player selects position in prompt
When running the game loop in a while loop the rendering didn't
happen until the end of the while loop, only the events were dispatched
and stored in the event loop queu to later be rendered when the while was
terminated.

To combat this I declared asyncGameLoop() which is a recurssive function, that
calls itself until the game is finished. This function is declared inside the
module Game within the start() so it has access to the player objects.