Handle case when user clicks in a board cell that is already filled
When a user clicks in a cell that's already filled a while loop will start again until
The position selected is empty and the marker is placed. When the user clicks on a board cell
that's filled a message will be displayed above the board saying that the position selected
is already filled and they need to try again.