Refactor initial values of board and renderBoardState()
Instead of populating the initial gameState array with 0
we just populate it with empty strings.

Whenever a click event is triggered in the board cells we run
renderBoardState(). This time this function just renders the content
that's inside of the boardState array. We don't need to check if it's
different from 0 just load all the values from the array.