Fix reseting behaviour when clicking reset button
Before when the reset button was clicked, only the board was erased,
but the game didn't start again and everytime you clicked on the start
button after finishing a game the markers didn't render properly.

Now when you click on the reset another game is initialized with the same
players introduced.